update and fix 2019-3-8
update and fix 2019-3-8

+++ Java/Loon-Neo/src/loon/utils/processes/RealtimeProcessManager.java
@@ -63,19 +63,25 @@ public class RealtimeProcessManager implements RealtimeProcessEvent, LRelease {
 				toBeUpdated = new SortedList<GameProcess>(this.processes);
 			}
 			final SortedList<GameProcess> deadProcesses = new SortedList<GameProcess>();
-			for (LIterator<GameProcess> it = toBeUpdated.listIterator(); it
-					.hasNext();) {
+			for (LIterator<GameProcess> it = toBeUpdated.listIterator(); it.hasNext();) {
 				GameProcess realtimeProcess = it.next();
-				realtimeProcess.tick(time);
-				if (realtimeProcess.isDead()) {
-					deadProcesses.add(realtimeProcess);
+				if (realtimeProcess != null) {
+					synchronized (realtimeProcess) {
+						realtimeProcess.tick(time);
+						if (realtimeProcess.isDead()) {
+							deadProcesses.add(realtimeProcess);
+						}
+					}
 				}
 			}
 			if (deadProcesses.size > 0) {
-				for (LIterator<GameProcess> it = deadProcesses.listIterator(); it
-						.hasNext();) {
+				for (LIterator<GameProcess> it = deadProcesses.listIterator(); it.hasNext();) {
 					GameProcess realtimeProcess = it.next();
-					realtimeProcess.finish();
+					if (realtimeProcess != null) {
+						synchronized (realtimeProcess) {
+							realtimeProcess.finish();
+						}
+					}
 				}
 				synchronized (this.processes) {
 					this.processes.removeAll(deadProcesses);
@@ -87,8 +93,7 @@ public class RealtimeProcessManager implements RealtimeProcessEvent, LRelease {
 	public GameProcess find(String id) {
 		if (processes != null && processes.size > 0) {
 			synchronized (this.processes) {
-				for (LIterator<GameProcess> it = processes.listIterator(); it
-						.hasNext();) {
+				for (LIterator<GameProcess> it = processes.listIterator(); it.hasNext();) {
 					GameProcess p = it.next();
 					if (p.getId() == id || p.getId().equals(id)) {
 						return p;
@@ -102,13 +107,14 @@ public class RealtimeProcessManager implements RealtimeProcessEvent, LRelease {
 	public void delete(String id) {
 		if (processes != null && processes.size > 0) {
 			synchronized (this.processes) {
-				final TArray<GameProcess> ps = new TArray<GameProcess>(
-						processes);
+				final TArray<GameProcess> ps = new TArray<GameProcess>(processes);
 				for (int i = 0; i < ps.size; i++) {
 					GameProcess p = ps.get(i);
-					if (p.getId() == id || p.getId().equals(id)) {
-						p.kill();
-						processes.remove(p);
+					if (p != null) {
+						if (p.getId() == id || p.getId().equals(id)) {
+							p.kill();
+							processes.remove(p);
+						}
 					}
 				}
 			}
@@ -118,20 +124,21 @@ public class RealtimeProcessManager implements RealtimeProcessEvent, LRelease {
 	public void deleteIndex(String id) {
 		if (processes != null && processes.size > 0) {
 			synchronized (this.processes) {
-				final TArray<GameProcess> ps = new TArray<GameProcess>(
-						processes);
+				final TArray<GameProcess> ps = new TArray<GameProcess>(processes);
 				for (int i = 0; i < ps.size; i++) {
 					GameProcess p = ps.get(i);
-					if (p.getId() == id || p.getId().indexOf(id) != -1) {
-						p.kill();
-						processes.remove(p);
+					if (p != null) {
+						if (p.getId() == id || p.getId().indexOf(id) != -1) {
+							p.kill();
+							processes.remove(p);
+						}
 					}
 				}
 			}
 		}
 	}
-	
-	public void dispose(){
+
+	public void dispose() {
 		close();
 	}
 
@@ -139,11 +146,14 @@ public class RealtimeProcessManager implements RealtimeProcessEvent, LRelease {
 	public void close() {
 		if (processes != null && processes.size > 0) {
 			synchronized (this.processes) {
-				final TArray<GameProcess> ps = new TArray<GameProcess>(
-						processes);
+				final TArray<GameProcess> ps = new TArray<GameProcess>(processes);
 				for (int i = 0; i < ps.size; i++) {
 					GameProcess p = ps.get(i);
-					p.finish();
+					if (p != null) {
+						synchronized (p) {
+							p.finish();
+						}
+					}
 				}
 				processes.clear();
 			}
